debugging the rebase and externalizing the TextRender
- Avoids the #include "Viewer.h" in Scene_polyhedron_item.h"
