+++ Polyhedron/demo/Polyhedron/CMakeLists.txt
@@ -161,12 +161,20 @@ if(CGAL_Qt5_FOUND AND Qt5_FOUND AND OPENGL_FOUND AND QGLVIEWER_FOUND)
     DOC "Path to CGAL/Three/Scene_item.h")
   hide_variable(CGAL_THREE_SCENE_ITEM_H_PATH)
 
+  find_path(CGAL_THREE_TEXTRENDERER_H_PATH
+    NAME CGAL/Three/TextRenderer.h
+    PATHS ${CGAL_INCLUDE_DIRS}
+    NO_DEFAULT_PATH
+    DOC "Path to CGAL/Three/TextRenderer.h")
+  hide_variable(CGAL_THREE_TEXTRENDERER_H_PATH)
+
   find_path(CGAL_THREE_SCENE_GROUP_ITEM_H_PATH
     NAME CGAL/Three/Scene_group_item.h
     PATHS ${CGAL_INCLUDE_DIRS}
     NO_DEFAULT_PATH
     DOC "Path to CGAL/Three/Scene_group_item.h")
   hide_variable(CGAL_THREE_SCENE_GROUP_ITEM_H_PATH)
+
 
   if(CGAL_THREE_SCENE_ITEM_H_PATH)
     qt5_generate_moc( "${CGAL_THREE_SCENE_ITEM_H_PATH}/CGAL/Three/Scene_item.h"
@@ -182,7 +190,12 @@ if(CGAL_Qt5_FOUND AND Qt5_FOUND AND OPENGL_FOUND AND QGLVIEWER_FOUND)
     message(FATAL_ERROR "Cannot find <CGAL/Three/Scene_group_item.h>")
   endif()
 
-
+  if(CGAL_THREE_TEXTRENDERER_H_PATH)
+    qt5_generate_moc( "${CGAL_THREE_TEXTRENDERER_H_PATH}/CGAL/Three/TextRenderer.h"
+      "${CMAKE_CURRENT_BINARY_DIR}/TextRenderer_moc.cpp" )
+  else()
+    message(FATAL_ERROR "Cannot find <CGAL/Three/TextRenderer.h>")
+  endif()
   # AUXILIARY LIBRARIES
 
   # put s (which are shared libraries) at the same location as
@@ -198,6 +211,8 @@ if(CGAL_Qt5_FOUND AND Qt5_FOUND AND OPENGL_FOUND AND QGLVIEWER_FOUND)
     Scene_item.cpp
     Scene_group_item.cpp
     Scene_group_item_moc.cpp
+    TextRenderer.cpp
+    TextRenderer_moc.cpp
     Polyhedron_demo_plugin_helper.cpp)
     qt5_use_modules(demo_framework OpenGL Gui Widgets Script Xml)
